Change the directory for the Ghidra project
There's no guarantee that /usr/local will be writable (in fact, on both
Linux and macOS, I had permission errors and Ghidra was unable to lock
the project directory). Therefore, store the project in the temporary
directory instead (this is OS-dependent and is retrieved via the
`java.io.tmpdir` system property).