Fix PATH variable conflict in Makefile
Using `PATH` as the variable name for the path to a firmware file
conflicts with the shell's `PATH`, which indicates where executable
files can be found. This results in an error because the java executable
can't be found when `PATH` is overwritten. Hence, the Makefile now uses
`FW_PATH` (for "firmware path") instead.